Add the expand/collapse feature
This allows the user to maximize or minimize the code output.

This commit also removes the hover effect on the sections, and the
scroll to the bottom feature.

The first was removed because it conflited with the new feature,
and the second because it was not needed anymore, not tha the user
can maximize the output.